+++ lib/game.h
@@ -1,0 +1,5 @@
+#include "engine.h"
+#include "character.h"
+#include "bullet.h"
+#include "enemy.h"
+
